meson: Move all compilation and generation to src/
In the process, remove headers files from the list of sources. This is
not necessary with Meson.

The move is required for generating enum files with gnome.mkenums() and
for compiling resources with gnome.compile_resources()

Also set meson_version to set a lower bound on compatibility, and use
conf.set_quoted which quotes strings for us for use in C headers.